+++ wpqs.cpp
@@ -1,7 +1,7 @@
 #include "wpqs.h"
 #include <iostream>
 
-void Wpqs::addJob(int job){
+void Wpqs::addJob(unsigned long job){
 	priority_jobs.push(job);
 }
 
@@ -10,7 +10,7 @@ void Wpqs::schedule(){
 		Request head = priority_jobs.top();
 		if (head.size <= resources.size()){
 			std::queue<Request> _resources;
-			for (int i = 0; i < head.size; i++){
+			for (unsigned long i = 0; i < head.size; i++){
 				_resources.push(resources.front());
 				resources.pop();
 			}
